% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_wrappers.R
\name{bm_2023_bilateral_pure_all}
\alias{bm_2023_bilateral_pure_all}
\title{BM_2023 pure bilateral (/sr) decomposition for all pairs}
\usage{
bm_2023_bilateral_pure_all(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame of pure bilateral decomposition for all pairs.
}
\description{
BM_2023 pure bilateral (/sr) decomposition for all pairs
}
