% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_bilateral.R
\name{bm_2023_bilateral_source}
\alias{bm_2023_bilateral_source}
\title{BM_2023 source-based bilateral decomposition of e_sr}
\usage{
bm_2023_bilateral_source(io, s, r)
}
\description{
BM_2023 source-based bilateral decomposition of e_sr
}
