% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_exporter.R
\name{bm_2023_exporter_total}
\alias{bm_2023_exporter_total}
\title{BM_2023 exporter-perspective decomposition of total exports of s}
\usage{
bm_2023_exporter_total(io, s)
}
\arguments{
\item{io}{bm_io object}

\item{s}{exporting country (index or code, e.g. 1 or "China")}
}
\value{
data.frame with one row for country s
}
\description{
Implements Eq. (4) at the country level:
u_N × e_s* = DVA_s* + DDC_s* + FVA_s* + FDC_s*
}
