% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output.R
\name{bm_2025_output_components}
\alias{bm_2025_output_components}
\title{BM_2025 output-based GVC components by exporter}
\usage{
bm_2025_output_components(io)
}
\arguments{
\item{io}{bm_io object}
}
\value{
data.frame with one row per country.
}
\description{
Computes output-based PF/TS/PB GVC components for each country s.
}
