% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output.R
\name{bm_2025_output_measures}
\alias{bm_2025_output_measures}
\title{BM_2025 output-based GVC participation indicators}
\usage{
bm_2025_output_measures(io)
}
\arguments{
\item{io}{bm_io object}
}
\value{
data.frame with exporter-level output-based indicators.
}
\description{
For each country s:
\itemize{
\item share_GVC_output = GVC_X / X_total
\item share_PF_output  = GVC_PF_X / GVC_X
\item share_TS_output  = GVC_TS_X / GVC_X
\item share_PB_output  = GVC_PB_X / GVC_X
\item forward_output   = (GVC_PF_X - GVC_PB_X) / GVC_X
}
}
