% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output_sector.R
\name{bm_2025_output_measures_sector}
\alias{bm_2025_output_measures_sector}
\title{BM 2025 output participation measures by country and sector}
\usage{
bm_2025_output_measures_sector(io)
}
\description{
BM 2025 output participation measures by country and sector
}
