% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_trade.R
\name{bm_2025_trade_exporter}
\alias{bm_2025_trade_exporter}
\title{BM_2025 exporter-level GVC trade totals}
\usage{
bm_2025_trade_exporter(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame of exporter totals.
}
\description{
BM_2025 exporter-level GVC trade totals
}
