% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_trade.R
\name{bm_2025_tripartite_trade_all}
\alias{bm_2025_tripartite_trade_all}
\title{BM_2025 tripartite GVC trade decomposition for all pairs}
\usage{
bm_2025_tripartite_trade_all(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame for all pairs.
}
\description{
BM_2025 tripartite GVC trade decomposition for all pairs
}
