% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_tripartite_trade.R
\name{bm_davax_sr}
\alias{bm_davax_sr}
\title{DAVAX_sr: Domestic value added absorbed in final demand in r}
\usage{
bm_davax_sr(io, s, r)
}
\arguments{
\item{io}{A \code{bm_io} object created by \code{bm_build_io()}.}

\item{s}{Exporting country (index or name).}

\item{r}{Importing country (index or name).}
}
\value{
Numeric scalar, DAVAX_sr.
}
\description{
Implements the DAVAX component for exports from s to r in BM_2025
tripartite GVC trade decomposition.
}
\details{
It corresponds (in your notation) to:
\eqn{DAVAX_{sr} = v_s L_{ss} ( y_{sr} + A_{sr} L_{rr} y_{rr} )}.
}
