% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_get_e_sr}
\alias{bm_get_e_sr}
\title{Exports from s to r by sector (intermediate + final)}
\usage{
bm_get_e_sr(io, exporter, importer)
}
\arguments{
\item{io}{bm_io object}

\item{exporter}{exporter country (index or name)}

\item{importer}{importer country (index or name)}
}
\value{
Numeric vector of length N with sectoral exports s -> r.
}
\description{
Computes a length-N vector of exports from exporter \eqn{s} to importer \eqn{r},
summing intermediate and final exports:
\deqn{ e_{sr} = Z_{s,r} \mathbf{u}_N + Y_{s,r} }
}
\details{
\code{exporter} and \code{importer} can be numeric indices or country names.
}
