% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_gvc_measures.R
\name{bm_gvc_measures_all}
\alias{bm_gvc_measures_all}
\title{Combined trade- and output-based GVC measures}
\usage{
bm_gvc_measures_all(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
A named list with components \code{trade} and \code{output}.
}
\description{
Returns a list with:
\itemize{
\item trade  : trade-based decomposition + indicators
\item output : output-based decomposition + indicators
}
}
