% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_output_gvc.R
\name{bm_gvc_output_all}
\alias{bm_gvc_output_all}
\title{Output-based GVC decomposition for all countries}
\usage{
bm_gvc_output_all(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
A \code{data.frame} with one row per country.
}
\description{
Output-based GVC decomposition for all countries
}
