% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_output_gvc.R
\name{bm_gvc_pf_output}
\alias{bm_gvc_pf_output}
\title{Pure-forward GVC-related output (country level)}
\usage{
bm_gvc_pf_output(io, s, Xexp_list = NULL)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Country (index or name).}

\item{Xexp_list}{Optional precomputed list of export-related output
(as from \code{bm_build_Xexp_list(io)}). If \code{NULL}, it is
built internally.}
}
\value{
Numeric scalar: pure-forward GVC-related output of s.
}
\description{
For a given country s, this function computes the pure-forward
GVC-related output, i.e. domestic value added of s embodied in r's
export-related output to third countries (forward chains).
}
