% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_output_gvc.R
\name{bm_gvc_ts_import_output}
\alias{bm_gvc_ts_import_output}
\title{Two-sided GVC-related output from imported inputs}
\usage{
bm_gvc_ts_import_output(io, s, gvc_pb_s)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Country (index or name).}

\item{gvc_pb_s}{Pre-computed pure-backward GVC output of s
(from \code{bm_gvc_pb_output}).}
}
\value{
Numeric scalar: imported two-sided GVC-related output of s.
}
\description{
Imported two-sided GVC-related output of s is defined as all imported
inputs used in s's output, net of those absorbed in purely domestic
chains, net of the pure-backward part.
}
