% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_tripartite_trade.R
\name{bm_tripartite_trade_sr}
\alias{bm_tripartite_trade_sr}
\title{Tripartite GVC trade decomposition for a bilateral pair (s,r)}
\usage{
bm_tripartite_trade_sr(io, s, r)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Exporting country (index or name).}

\item{r}{Importing country (index or name).}
}
\value{
A one-row \code{data.frame} with all components.
}
\description{
For exports from s to r, this function returns:
\itemize{
\item E_sr: gross exports (sum of e_sr)
\item DAVAX_sr: domestic value added that crosses only one border and is absorbed in r's final demand
\item GVC_sr: GVC-related part of E_sr, GVC_sr = E_sr - DAVAX_sr
\item GVC_PF: pure-forward GVC trade (residual)
\item GVC_TS: two-sided GVC trade
\item GVC_PB: pure-backward GVC trade
}
}
\details{
In your notation:
\deqn{E_{sr} = DAVAX_{sr} + GVC_{sr}}
\deqn{GVC_{sr} = GVC^{PF}_{sr} + GVC^{TS}_{sr} + GVC^{PB}_{sr}}.
}
