% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_bilateral.R
\name{bm_2023_bilateral_pure}
\alias{bm_2023_bilateral_pure}
\title{BM_2023 pure bilateral decomposition of exports from s to r}
\usage{
bm_2023_bilateral_pure(io, s, r)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Exporter country (name or index).}

\item{r}{Importer country (name or index).}
}
\value{
A data frame with the pure bilateral value-added decomposition.
}
\description{
BM_2023 pure bilateral decomposition of exports from s to r
}
