% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_exporter.R
\name{bm_2023_exporter_total}
\alias{bm_2023_exporter_total}
\title{BM_2023 exporter-perspective decomposition of total exports of s}
\usage{
bm_2023_exporter_total(io, s)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Exporter country (name or index).}
}
\value{
A data frame with the exporter-total decomposition.
}
\description{
BM_2023 exporter-perspective decomposition of total exports of s
}
