% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2023_wrappers.R
\name{bm_2023_exporter_total_all}
\alias{bm_2023_exporter_total_all}
\title{BM_2023 exporter totals for all countries}
\usage{
bm_2023_exporter_total_all(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame of exporter totals for all countries.
}
\description{
BM_2023 exporter totals for all countries
}
