% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output.R
\name{bm_2025_output_components}
\alias{bm_2025_output_components}
\title{BM_2025 output-based GVC components by exporter}
\usage{
bm_2025_output_components(io)
}
\description{
BM_2025 output-based GVC components by exporter
}
