% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output.R
\name{bm_2025_output_measures}
\alias{bm_2025_output_measures}
\title{BM_2025 output-based GVC participation indicators}
\usage{
bm_2025_output_measures(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame with output-based GVC participation measures.
}
\description{
BM_2025 output-based GVC participation indicators
}
