% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_output.R
\name{bm_2025_output_measures}
\alias{bm_2025_output_measures}
\title{BM_2025 output-based GVC participation indicators}
\usage{
bm_2025_output_measures(io)
}
\description{
BM_2025 output-based GVC participation indicators
}
