% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_trade.R
\name{bm_2025_trade_measures}
\alias{bm_2025_trade_measures}
\title{BM_2025 trade-based GVC participation indicators}
\usage{
bm_2025_trade_measures(io)
}
\arguments{
\item{io}{A \code{bm_io} object.}
}
\value{
Data frame of trade-based indicators.
}
\description{
BM_2025 trade-based GVC participation indicators
}
