% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_2025_trade.R
\name{bm_2025_tripartite_trade}
\alias{bm_2025_tripartite_trade}
\title{BM_2025 tripartite GVC trade decomposition for one pair (s,r)}
\usage{
bm_2025_tripartite_trade(io, s, r)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Exporter country (name or index).}

\item{r}{Importer country (name or index).}
}
\value{
Data frame for the pair (s,r).
}
\description{
BM_2025 tripartite GVC trade decomposition for one pair (s,r)
}
