% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_build_io}
\alias{bm_build_io}
\title{Build a bm_io object from IO table blocks}
\usage{
bm_build_io(Z, Y, VA, X, countries, sectors)
}
\arguments{
\item{Z}{Intermediate demand matrix (GN x GN), rows = suppliers, cols = users.}

\item{Y}{Final demand matrix (GN x G), columns = destination countries.}

\item{VA}{Value added vector (length GN), in same industry order as Z rows.}

\item{X}{Output vector (length GN), in same industry order as Z rows.}

\item{countries}{Character vector of country names, length G.}

\item{sectors}{Character vector of sector names, length N.}
}
\value{
An object of class \code{"bm_io"} containing
\itemize{
\item \code{Z, Y, VA, X}
\item \code{countries, sectors}
\item \code{G, N, GN}
\item \code{A}: technical coefficients (GN x GN)
\item \code{B}: Leontief inverse (GN x GN)
\item \code{v}: value-added coefficients (length GN)
}
}
\description{
Build a bm_io object from IO table blocks
}
