% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_build_io}
\alias{bm_build_io}
\title{Build a bm_io object from IO table blocks}
\usage{
bm_build_io(Z, Y, VA, X, countries, sectors)
}
\arguments{
\item{Z}{Intermediate demand matrix (GN x GN).}

\item{Y}{Final demand matrix. Can be (GN x G) OR (GN x (G * FD_categories)).}

\item{VA}{Value added. Can be a vector (length GN) or matrix (Rows x GN).}

\item{X}{Output vector (length GN).}

\item{countries}{Character vector of country names/codes (length G).}

\item{sectors}{Character vector of sector names/codes (length N).}
}
\value{
An object of class \code{"bm_io"}.
}
\description{
Build a bm_io object from IO table blocks
}
