% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_build_io}
\alias{bm_build_io}
\title{Build a BM-style IO object}
\usage{
bm_build_io(Z, Y, VA, X, countries, sectors)
}
\arguments{
\item{Z}{GN x GN intermediate-use matrix}

\item{Y}{GN x G  final demand by destination country}

\item{VA}{length-GN vector of value added}

\item{X}{length-GN vector of gross output}

\item{countries}{character vector of country codes/names (length G)}

\item{sectors}{character vector of sector codes/names (length N)}
}
\value{
A list of IO objects and metadata used by bmGVC
}
\description{
Build a BM-style IO object
}
