% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_indices.R
\name{bm_country_id}
\alias{bm_country_id}
\title{Resolve a country (name or index) to numeric index 1..G}
\usage{
bm_country_id(io, country)
}
\arguments{
\item{io}{A bm_io object containing the MRIO data.}

\item{country}{Country code or index used to select a country in \code{io}.}
}
\description{
Resolve a country (name or index) to numeric index 1..G
}
