% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_get_e_sr}
\alias{bm_get_e_sr}
\title{Bilateral exports e_sr by sector}
\usage{
bm_get_e_sr(io, s, r)
}
\arguments{
\item{io}{bm_io object.}

\item{s}{Exporter index (1..G).}

\item{r}{Importer index (1..G).}
}
\value{
Numeric vector of length N with exports from s to r by sector.
}
\description{
Total exports from country s to country r by sector, including
both intermediate and final goods.
}
