% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_get_e_star}
\alias{bm_get_e_star}
\title{Country-total exports e_s* by sector}
\usage{
bm_get_e_star(io, s)
}
\arguments{
\item{io}{bm_io object.}

\item{s}{Exporter index (1..G).}
}
\value{
Numeric vector of length N with exports from s to the world.
}
\description{
Total exports from country s to all foreign destinations,
decomposed by sector.
}
