% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_io.R
\name{bm_get_e_star}
\alias{bm_get_e_star}
\title{Total exports of s to all foreign destinations}
\usage{
bm_get_e_star(io, exporter)
}
\arguments{
\item{io}{bm_io object.}

\item{exporter}{Exporter country (name or index).}
}
\value{
Numeric vector of total exports.
}
\description{
Total exports of s to all foreign destinations
}
