% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_tripartite_trade.R
\name{bm_pb_sr}
\alias{bm_pb_sr}
\title{PB_sr: Pure-backward GVC trade (import content used in r's final demand)}
\usage{
bm_pb_sr(io, s, r)
}
\arguments{
\item{io}{A \code{bm_io} object.}

\item{s}{Exporting country (index or name).}

\item{r}{Importing country (index or name).}
}
\value{
Numeric scalar, PB_sr.
}
\description{
Pure-backward GVC trade is the import content of production in s
that serves r's final demand of s and r goods.
}
\details{
In your notation:
\itemize{
\item q_s(sr) = L_ss ( y_{sr} + A_{sr} L_{rr} y_{rr} )
\item PB_sr = Σ_{t≠s} ( u_N A_{ts} q_s(sr) ).
}
}
