% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_indices.R
\name{bm_sector_id}
\alias{bm_sector_id}
\title{Resolve a sector (name or index) to numeric index 1..N}
\usage{
bm_sector_id(io, sector)
}
\description{
Resolve a sector (name or index) to numeric index 1..N
}
